Added in place arithmetic to lexer
The lexer now includes most of the in place arithmetic operators I would care to support. There are still a few bit-wise logic operators to come. There was this strange bug I noticed where it would take an input like 'a //= b' and spit out
[IDENTIFIER, a], [DBL_FWD_SLASH, //], [FLOOR_DIVIDE_EQU, //=], [IDENTIFIER b]

Stranger still, it doesn't do it with **=, which I totally expected it to do while I was writing the changes. I fixed it by throwing in a
del tokens[-2] after I reassigned the last token to the correct type and value but it's highly unusual and a likely sign of problems that could emerge in later updates.

I've also made a few minor corrections to the grammar with more to come.