Updated the syntax and completed? the lexer
The lexer now supports all the keywords and symbols supported in the language. I also changed the getIdentifierOrKeyword method to split on the dots, remove empty strings, then convert the result into a tuple to make parsing easier. The syntax may still require a variety of improvements but all cases where member field references should be allowed are now defined accordingly and the syntax definition is more or less how I want it. Parsing expressions is going to be a challenge.