Fixed indexing information used for error tracing
I corrected the indexing information used for tracing syntax errors during parsing and other compilation errors. I also fixed a bug that occurs when performing lexical analysis on block strings. That is, """This string can
span multiple lines."""
Having realized I hadn't added the syntax definition for block strings to the syntax def, I proceeded to make that correction to the grammar and make a correction to the StringBody terminal.