Changed the lexical analysis of comments
Previously comments were seen as newlines of the \n variety. I realized that this was problematic as it made it difficult to insert comments on their own line within code blocks without the comment being seen as the end of the block by the parser.
I also changed the error message that appears when 'anIdentifier.' or . is seen so as to make it clear to the programmer that spaces aren't allowed in member field references. That is Expr.anIdentifier, not expr. anIdentifier or expr . anIdentifier.